fix page refresh redirect
